requirements.txt
Pillow
psutil
pyshark
pandas
fpdf
